Extract IngestBatcherTestBase
