+++ java/ingest/ingest-batch-core/src/test/java/sleeper/ingest/batcher/IngestBatcherTest.java
@@ -16,53 +16,23 @@
 
 package sleeper.ingest.batcher;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import sleeper.configuration.properties.InstanceProperties;
-import sleeper.configuration.properties.table.FixedTablePropertiesProvider;
 import sleeper.configuration.properties.table.TableProperties;
-import sleeper.configuration.properties.validation.BatchIngestMode;
-import sleeper.ingest.job.IngestJob;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
-import java.util.function.Supplier;
-import java.util.stream.IntStream;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.configuration.properties.InstancePropertiesTestHelper.createTestInstanceProperties;
-import static sleeper.configuration.properties.SystemDefinedInstanceProperty.INGEST_JOB_QUEUE_URL;
-import static sleeper.configuration.properties.table.TablePropertiesTestHelper.createTestTableProperties;
-import static sleeper.configuration.properties.table.TableProperty.INGEST_BATCHER_INGEST_MODE;
 import static sleeper.configuration.properties.table.TableProperty.INGEST_BATCHER_MAX_JOB_FILES;
 import static sleeper.configuration.properties.table.TableProperty.INGEST_BATCHER_MAX_JOB_SIZE;
 import static sleeper.configuration.properties.table.TableProperty.INGEST_BATCHER_MIN_JOB_FILES;
 import static sleeper.configuration.properties.table.TableProperty.INGEST_BATCHER_MIN_JOB_SIZE;
-import static sleeper.configuration.properties.table.TableProperty.TABLE_NAME;
-import static sleeper.core.schema.SchemaTestHelper.schemaWithKey;
-import static sleeper.ingest.batcher.FileIngestRequestTestHelper.DEFAULT_TABLE_NAME;
 import static sleeper.ingest.batcher.FileIngestRequestTestHelper.onJob;
 
-class IngestBatcherTest {
-    private final InstanceProperties instanceProperties = createTestInstanceProperties();
-    private final TableProperties tableProperties = createTableProperties(DEFAULT_TABLE_NAME);
-    private final IngestBatcherStateStore store = new IngestBatcherStateStoreInMemory();
-    private final IngestBatcherQueuesInMemory queues = new IngestBatcherQueuesInMemory();
-    private final FileIngestRequestTestHelper requests = new FileIngestRequestTestHelper();
-
-    @BeforeEach
-    void setUp() {
-        instanceProperties.set(INGEST_JOB_QUEUE_URL, "test-ingest-queue-url");
-    }
-
-    private Map<String, List<Object>> queueMessages(IngestJob... jobs) {
-        return Map.of("test-ingest-queue-url", List.of(jobs));
-    }
+class IngestBatcherTest extends IngestBatcherTestBase {
 
     @Nested
     @DisplayName("Batch with minimum file count")
@@ -555,59 +525,5 @@ class IngestBatcherTest {
                             jobWithFiles("test-job-id-3",
                                     "test-bucket/test-4.parquet")));
         }
-    }
-
-    private TableProperties createTableProperties(String tableName) {
-        TableProperties properties = createTestTableProperties(instanceProperties, schemaWithKey("key"));
-        properties.set(INGEST_BATCHER_INGEST_MODE, BatchIngestMode.STANDARD_INGEST.toString());
-        properties.set(INGEST_BATCHER_MIN_JOB_SIZE, "0");
-        properties.set(INGEST_BATCHER_MIN_JOB_FILES, "0");
-        properties.set(TABLE_NAME, tableName);
-        return properties;
-    }
-
-    private IngestJob jobWithFiles(String jobId, String... files) {
-        return IngestJob.builder()
-                .files(files)
-                .tableName(DEFAULT_TABLE_NAME)
-                .id(jobId)
-                .build();
-    }
-
-    private FileIngestRequest addFileToStore(String pathToFile) {
-        return addFileToStore(ingestRequest()
-                .pathToFile(pathToFile).build());
-    }
-
-    private FileIngestRequest.Builder ingestRequest() {
-        return requests.fileRequest();
-    }
-
-    private FileIngestRequest addFileToStore(FileIngestRequest request) {
-        store.addFile(request);
-        return request;
-    }
-
-    private void batchFilesWithJobIds(String... jobIds) {
-        batchFilesWithTablesAndJobIds(List.of(tableProperties), List.of(jobIds));
-    }
-
-    private void batchFilesWithTablesAndJobIds(List<TableProperties> tables, List<String> jobIds) {
-        IngestBatcher.builder()
-                .instanceProperties(instanceProperties)
-                .tablePropertiesProvider(new FixedTablePropertiesProvider(tables))
-                .jobIdSupplier(jobIdSupplier(jobIds))
-                .store(store).queueClient(queues)
-                .build().batchFiles();
-    }
-
-    private static Supplier<String> jobIdSupplier(List<String> jobIds) {
-        return Stream.concat(jobIds.stream(), infiniteIdsForUnexpectedJobs())
-                .iterator()::next;
-    }
-
-    private static Stream<String> infiniteIdsForUnexpectedJobs() {
-        return IntStream.iterate(1, n -> n + 1)
-                .mapToObj(num -> "unexpected-job-" + num);
     }
 }
